packages/fabricatio-agent/pyproject.toml
[project]
name = "fabricatio-agent"
version = "0.1.4-dev12"
description = "An extension of fabricatio"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Whth", email = "zettainspector@foxmail.com" }
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Implementation :: CPython",
    "Typing :: Typed",
]
requires-python = ">=3.12,<3.15"
dependencies = [
    "fabricatio-core",
    "fabricatio-digest",
    "fabricatio-memory",
    "fabricatio-improve",
    "fabricatio-rule",
    "fabricatio-judge",
    "fabricatio-capabilities",
    "fabricatio-diff",
    "fabricatio-thinking",
    "fabricatio-question",
    "fabricatio-tool",
    "fabricatio-team",
    "fabricatio-capable",
    "fabricatio-checkpoint"
]

[dependency-groups]
dev = [
    "pytest",
]


[project.optional-dependencies]
cli = [
    "typer-slim[standard]>=0.15.2",
]



[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[tool.maturin]
python-source = "python"
module-name = "fabricatio_agent.rust"

[tool.uv.sources]
fabricatio-digest = { workspace = true }

[[tool.uv.index]]
url = "https://mirrors.huaweicloud.com/repository/pypi/simple"

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[project.scripts]
falcon = "fabricatio_agent.cli:app"

[project.urls]
Homepage = "https://github.com/Whth/fabricatio"
Repository = "https://github.com/Whth/fabricatio"
Issues = "https://github.com/Whth/fabricatio/issues"
